refactor: keep the core/widgets separation
