+++ docs/src/demos/Atoms/Button/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
-import Button from 'Button'
-import SplitButton from 'SplitButton'
-import DropdownButton from 'DropdownButton'
-import Icon from 'Icon'
+import Button from 'core/Button'
+import SplitButton from 'core/SplitButton'
+import DropdownButton from 'core/DropdownButton'
+import Icon from 'core/Icon'
 
 import { Row, Divider, Heading } from '../../../helpers'
