+++ docs/src/demos/Atoms/Card/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Card from 'Card'
+import Card from 'core/Card'
 import { Col, Divider } from '../../../helpers'
 
 export class CardDemo extends React.Component {
