+++ docs/src/demos/Atoms/Input/Text/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import InputField from 'InputField'
+import InputField from 'core/InputField'
 
 import { Text, Row, Col, Divider } from '../../../../helpers'
