+++ docs/src/demos/Atoms/Menu/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Menu from 'Menu'
-import Icon from 'Icon'
+import Menu from 'core/Menu'
+import Icon from 'core/Icon'
 import { Text, Col, Divider, Spacer } from '../../../helpers'
 
 const list = []
