+++ docs/src/helpers.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Divider as CoreDivider } from 'Divider'
+import { Divider as CoreDivider } from 'core/Divider'
 
 export const Title = ({ children }) => <h1 className="title">{children}</h1>
 export const Heading = ({ children }) => <h2 className="heading">{children}</h2>
