+++ src/core/Chip/index.js
