+++ src/core/CircularProgress/index.js
