+++ src/core/CircularProgress/styles.js
