+++ src/core/Divider/index.js
