+++ src/core/Icon/index.js
