+++ src/core/Icon/styles.js
