+++ src/core/LinearProgress/index.js
