+++ src/core/LinearProgress/styles.js
