+++ src/core/Logo/index.js
