+++ src/core/Switch/index.js
