+++ src/core/Switch/styles.js
