+++ src/index.js
@@ -1,2 +1,21 @@
-export * from './core'
-export * from './widgets'
+// Core
+export { default as Button } from './core/Button'
+export { default as Card } from './core/Card'
+export { default as Checkbox } from './core/Checkbox'
+export { default as Chip } from './core/Chip'
+export { default as CircularProgress } from './core/CircularProgress'
+export { default as Divider } from './core/Divider'
+export { default as DropdownButton } from './core/DropdownButton'
+export { default as Help } from './core/Help'
+export { default as Icon } from './core/Icon'
+export { default as InputField } from './core/InputField'
+export { default as LinearProgress } from './core/LinearProgress'
+export { default as Logo } from './core/Logo'
+export { default as Menu } from './core/Menu'
+export { default as Radio } from './core/Radio'
+export { default as SelectField } from './core/SelectField'
+export { default as SplitButton } from './core/SplitButton'
+export { default as Switch } from './core/Switch'
+
+// Widgets
+export { default as HeaderBar } from './widgets/HeaderBar'
